Restructured project layout
